refactor: Update import paths for use-taken-lecture hook
