--- app/type/lecture.ts
@@ -1,9 +1,0 @@
-export interface TakenLectrueInfo {
-  [index: string]: string | number;
-  id: number;
-  year: string;
-  semester: string;
-  lectureCode: string;
-  lectureName: string;
-  credit: number;
-}
